+++ stdlib/SparseArrays/test/sparse.jl
@@ -1587,11 +1587,11 @@ end
     # non-square:
     for m=1:4, n=2:4
         if m < 2 || n < 3
-            @test_throws DimensionMismatch spdiagm(0 => x,  1 => x, size=(m,n))
+            @test_throws DimensionMismatch spdiagm(m,n, 0 => x,  1 => x)
         else
             M = zeros(m,n)
             M[1:2,1:3] = [1 1 0; 0 1 1]
-            @test spdiagm(0 => x,  1 => x, size=(m,n)) == M
+            @test spdiagm(m,n, 0 => x,  1 => x) == M
         end
     end
 end