+++ NEWS.md
@@ -24,6 +24,8 @@ New library functions
 Standard library changes
 ------------------------
 
+* `diagm` and `spdiagm` now accept optional `m,n` initial arguments to specify a size ([#31654]).
+
 #### LinearAlgebra